+++ rollup.config.js
@@ -1,10 +1,11 @@
 export default {
+  name: 'leaflet.terminator',
   input: 'index.js',
   output: {
     file: 'L.Terminator.js',
     format: 'iife'
   },
-  external: ['L'],
+  external: ['leaflet'],
   globals: {
     'leaflet': 'L',
   }
